+++ src/matrix/ScalarMapping.c
@@ -1,6 +1,6 @@
 #include "ScalarMapping.h"
 
-ScalarMapping *Point2D_new(const double val, const size_t ind) {
+ScalarMapping *Point2D_new(const float val, const uint32_t ind) {
     ScalarMapping *point = malloc(sizeof(ScalarMapping));
     if (point != NULL) {
         point->value = val;
@@ -9,7 +9,7 @@ ScalarMapping *Point2D_new(const double val, const size_t ind) {
     return point;
 }
 
-void set_values(ScalarMapping *point, double val, size_t ind) {
+void set_values(ScalarMapping *point, float val, uint32_t ind) {
     if (point != NULL) {
         point->value = val;
         point->index = ind;
